requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.3
asyncio==3.4.3
httpcore==0.18.0
httpx==0.24.1
sqlite3
requests==2.31.0
beautifulsoup4==4.12.3
lxml==4.9.3
googletrans==4.0.0-rc1
pytz==2024.1
ccxt==2.3.1
yfinance==0.1.70
